// digital-first/apps/deck/src/app/components/deck/models/card-type-enum.ts
export enum CardType {
    Standard = 'Standard',
    Parent = 'Parent',
    Wiki = 'Wiki',
    Markdown = 'Markdown',
    Audio = 'Audio',
    Video = 'Video',
    Embed = 'Embed',
    Frame = 'Frame',
    Chart = 'Chart',
    Image = 'Image',
    Newtab = 'NewTab',
    Refiner = 'Refiner',
    BriefSummary = 'BriefSummary',
    StandardNonNav = 'StandardNonNav',
    StandardNonNavNoActions = 'StandardNonNavNoActions',
    SummaryLinksNoActions = 'SummaryLinksNoActions'
  }